Register the template converter as global ::Asciidoctor::Converter::TemplateConverter
