+++ lib/converter.js
@@ -7,10 +7,34 @@ function hash_to_native($hash) {
   return $hash.$$smap;
 }
 
-const C = outils.subclass({
-  '$initialize': function(templates){
-    debug("$initialize", templates);
-    this.templates = templates;
+/**
+  Load all templates from a directory and store them in a map.
+*/
+function load_templates(template_dir) {
+  const glob = require('glob');
+  const path = require('path');
+  const pug = require('pug');
+
+  const templates = new Map();
+
+  // XXX it is unfortunate to use a synchronous method here...
+  for(let file of glob.sync(path.join(template_dir, "*.pug"))) {
+    debug("Load", file);
+    const key = path.parse(file).name;
+    const template = pug.compileFile(file);
+
+    templates.set(key, template);
+  }
+
+  return templates;
+}
+
+// XXX This is still a little bit verbose...
+const opal_module = outils.const_get_qualified('::Asciidoctor::Converter');
+const opal_base_converter = outils.const_get_qualified('::Asciidoctor::Converter::Base');
+const C = outils.subclass(opal_module, opal_base_converter, 'TemplateConverter', {
+  '$initialize': function(backend, template_dirs, opts){
+    this.templates = load_templates(template_dirs);
   },
 
   '$handles?': function(name){
@@ -42,4 +66,4 @@ const C = outils.subclass({
   },
 });
 
-module.exports = C;
+return C;
