+++ lib/index.js
@@ -27,7 +27,6 @@ const PugConverter = require('./converter.js');
   // converted in a composite+template converter if
   // the 'template_dirs' argument is passed
   const inherited = outils.patch(Factory, '$create', function(backend, opts) {
-    debug("$create", backend, opts);
     const converter = inherited.call(this, backend, opts);
 
     if (opts.template_dirs) {
@@ -37,21 +36,4 @@ const PugConverter = require('./converter.js');
 
     return converter;
   });
-  debug("Patched", inherited);
 })();
-
-function load_templates(template_dir) {
-  const glob = require('glob');
-  const path = require('path');
-  const pug = require('pug');
-
-  const templates = new Map();
-  // XXX it is unfortunate to use a synchronous method here...
-  for(let file of glob.sync(path.join(template_dir, "*.pug"))) {
-    debug("Load", file);
-    const key = path.parse(file).name;
-    const template = pug.compileFile(file);
-
-    templates.set(key, template);
-  }
-}
